Introduce different behaviour for PUTs and GETs
Canonicalize according to schema is called both on PUTs and on
GET-requests, which are dealt with symmetrically, but they have
different requirements (For instance: on save, the last-modified date
should be set to the current date, but on get, the last-modified date
should be left as is in storage).

By specifying direction different behaviour is introduced.