Add default service type 'rest` in case `ams:serviceType` is missing
The synchronisation of dcat to overheid.nl is crashing on a missing
service type. Because the `rest` serviceType is by far the most used,
add that as an option.

Furthermore, as requested by Nico de Graaf, add an extra option (`gpkg`)
for service type.

As a warning a message is printed. Because this script is not
using a proper logger, a python `print()` is used.